[Web] Refactor the global layout folder
