+++ core/datacap-web/console-fe/src/views/common/http-code/NotFound.vue
