+++ core/datacap-web/console-fe/src/views/layout/common/components/LayoutContent.vue
