+++ core/datacap-web/console-fe/src/views/layout/common/components/LayoutFooter.vue
